[fix] 1. Add ive static library       2. Eliminate compilation warnings when open -Wall
Reviewed-on: https://gerrit-ai.sophgo.vip:8443/143397

+++ evaluation/eval_keypoint.cpp
@@ -44,7 +44,7 @@ void dump_kp_to_dir_txt(
     std::shared_ptr<ModelLandmarksInfo> lm =
         std::static_pointer_cast<ModelLandmarksInfo>(outs[h]);
     ofs << h;
-    for (int k = 0; k < lm->landmarks_x.size(); ++k)
+    for (size_t k = 0; k < lm->landmarks_x.size(); ++k)
       ofs << ' ' << lm->landmarks_x[k] << ' ' << lm->landmarks_y[k];
     ofs << '\n';
   }
