+++ src/components/nn/object_detection/yolov10.cpp
@@ -165,7 +165,7 @@ void YoloV10Detection::decodeBboxFeatureMap(int batch_idx, int stride,
     grid_logits = get_box_vals(p_box_float, num_anchor, anchor_idx,
                                num_box_channel_, qscale);
   } else {
-    LOGE("unsupported data type:%d\n", boxinfo.data_type);
+    LOGE("unsupported data type:%d\n", static_cast<int>(boxinfo.data_type));
     return;
   }
 
@@ -246,7 +246,7 @@ int32_t YoloV10Detection::outputParse(
                                 num_cls, j, cls_offset, cls_qscale, &max_logit,
                                 &max_logit_c);
         } else {
-          LOGE("unsupported data type:%d\n", classinfo.data_type);
+          LOGE("unsupported data type:%d\n", static_cast<int>(classinfo.data_type));
           assert(0);
         }
         if (max_logit < inverse_th) {
